Automatically lock Linux platform when possible
Since we started locking the specific platform in the lockfile, that has
created an annoying situation for users that don't develop on Linux.
They will create a lockfile on their machines, locking their local
platform, for example, darwin. But then that lockfile won't work
automatically when deploying to Heroku for example, because the lockfile
is frozen and the Linux platform is not included.

Since Linux is the main platform of our users (90% according to
ecosystem.rubytogether.org), it makes sense to special case it, and
automatically add it to the lockfile, even when generated for other
platform.

There's the chance though that resolving against two platforms (Linux +
the local platform) won't succeed while resolving for just the current
platform will. So, instead, we check whether the resolution found for
the current platform is also compatible with Linux, and only in that
case, we add it to the lockfile.

Also, to not risk any expected behavior for users of existing lockfiles,
we do this only when generating fresh lockfiles, and only on "equivalent
platforms", like darwin (Windows and Java have their own complexities,
and so are excluded).

With this change, we expect that MacOS users can bundle locally and
deploy to Heroku without needing to do anything special.