Add CHECKSUMS for each gem in lockfile
We lock the checksum for each resolved spec under a new CHECKSUMS
section in the lockfile.

The checksum is calculated from the SHA256 checksum of the cached gem.

If the locked spec does not resolve for the local platform, we preserve
the locked checksum, similar to how we preserve specs.